+++ packages/hooks/use-media-query/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @yamada-ui/use-media-query
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @yamada-ui/providers@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @yamada-ui/providers@1.0.1
+
 ## 1.0.0
 
 ### Major Changes